% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abdellaouirg.R
\docType{data}
\name{abdellaouietal}
\alias{abdellaouietal}
\title{Genetic correlations across psychiatric disorders}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
\href{https://pubmed.ncbi.nlm.nih.gov/33524898/}{}
}
\usage{
data(abdellaouirg)
}
\description{
Data from LDSC regression on 11 psychiatric phenotypes including substance use
}
\examples{
data(abdellaouirg)
abdellaouirg[,2]
}
\references{
Abdellaoui, A., Smit, D. J., van den Brink, W., Denys, D., & Verweij, K. J. (2021). Genomic relationships across psychiatric disorders including substance use disorders. Drug and alcohol dependence, 220, 108535.
(\href{https://pubmed.ncbi.nlm.nih.gov/33524898/}{PubMed})
}
\keyword{datasets}
