% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autorecode.R
\name{autorecode}
\alias{autorecode}
\title{Automatically recodes a correlation matrix so that the strongest correlation of each variable is always positive

This function transforms correlation matrices into venn diagrams. The shared surface area of circles corresponds to the shared variance (R squared) or to another metric (e.g. Pearson / spearman correlation). The algorithm is an approximation based on a quasi-Newton algorithm.}
\usage{
autorecode(cormat)
}
\arguments{
\item{cormat}{A (square) correlation matrix (or other n times n numeric matrix)}
}
\value{
c A correlation matrix with recoded values
}
\description{
Automatically recodes a correlation matrix so that the strongest correlation of each variable is always positive

This function transforms correlation matrices into venn diagrams. The shared surface area of circles corresponds to the shared variance (R squared) or to another metric (e.g. Pearson / spearman correlation). The algorithm is an approximation based on a quasi-Newton algorithm.
}
\references{
Martin Steppan (2022). corr2venn: Correlation to Venn diagramm. R package version 0.1.0.
}
\keyword{Correlation}
\keyword{plot,}
\keyword{psychometrics,}
\keyword{recoding}
