% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big5.R
\docType{data}
\name{big5}
\alias{big5}
\title{Answers to the Big Five Personality Test questionnaire (50 items)}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
\href{http://openpsychometrics.org/_rawdata/}{}
}
\usage{
data(big5)
}
\description{
Data from items from the International Personality Item Pool (N=19,719)
}
\examples{
data(big5)
hist(big5$E1)
}
\references{
Possible Questionnaire Format for Administering the 50-Item Set of IPIP Big-Five Factor Markers". International Personality Item Pool.
(\href{http://openpsychometrics.org/_rawdata/}{Link})
}
\keyword{datasets}
