% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeCormat.R
\name{completeCormat}
\alias{completeCormat}
\title{Completes a correlation matrix, if only upper or lower triangle are present and / or diagonal values are not 1}
\usage{
completeCormat(cormat)

completeCormat(c)
}
\arguments{
\item{cormat}{A (square) correlation matrix (or other n times n numeric matrix)}
}
\value{
c A complete correlation matrix
}
\description{
Completes a correlation matrix, if only upper or lower triangle are present and / or diagonal values are not 1
}
\examples{
c<-cor(mtcars)
c[upper.tri(c)]<-NA
c
}
\references{
Martin Steppan (2022). corr2venn: Correlation to Venn diagramm. R package version 0.1.0.
}
\keyword{data}
\keyword{handling}
