% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2venn.R
\name{cor2venn}
\alias{cor2venn}
\title{Correlation matrix to Venn diagram}
\usage{
cor2venn(cormat, Rsquared = TRUE, DistProp = TRUE, Recode = TRUE,
Coloring = c("PC1", "Mclust", "manual"))
}
\arguments{
\item{cormat}{A (square) correlation matrix (or other n times n numeric matrix)}

\item{Rsquared}{Shared surface areas correspond to R^2 (defaults to TRUE)}

\item{Recode}{Automatically recodes the correlation matrix so that the highest correlation of each variable is always positive (defaults to TRUE)}

\item{DistProp}{Distances between data points are proportional to correlations (positive correlation = close, negative correlation = distant; defaults to TRUE)}

\item{Coloring}{Coloring of circles. There are three options: "PC1"}
}
\value{
p A ggplot2 object showing the graphical approximation

xyr A three-column matrix providing x and y coordinates for all variables in the plot and the radius as depicted in the ggplot2 object

t CPU time required to find the optimized solution
}
\description{
This function transforms correlation matrices into venn diagrams. The shared surface area of circles corresponds to the shared variance (R squared) or to another metric (e.g. Pearson / spearman correlation). The algorithm is an approximation based on a quasi-Newton algorithm.
}
\examples{
cormat <- cor(mtcars[,2:13])

solve <- cor2venn(cormat)
plot(solve$p)
}
\references{
Steppan, M. cor2venn. A new method to visualize correlation matrices
}
\keyword{Correlation}
\keyword{plot,}
\keyword{psychometrics}
