% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2Vennplot.R
\name{cor2Vennplot}
\alias{cor2Vennplot}
\title{Plot a cor2Venn object using ggplot2}
\usage{
cor2venn(cormat, Rsquared = TRUE, cor2dist=c("0","1","2"), Recode = TRUE,
Coloring = c("PC1", "Mclust", "manual"))
}
\arguments{
\item{fillmode}{Coloring of circles. There are three options: "Eigen" = according to first eigenvector; "mclust" = according to cluster ananlysis of coordinates; "manual" = manual filling of nodes (use manualfill to provide a vector)}

\item{annotate}{True or false whether there is an annotation underneath the plot}

\item{manualcolors}{A vector to manually change the border color of each variable}

\item{manualfill}{A vector to manually change the fill color of variables}

\item{manualalphafill}{A vector to manually change the alpha fill of variables, defaults to 0.5}

\item{cor2Vennobj}{A cor2Venn object}
}
\value{
p A ggplot2 object showing the graphical approximation
}
\description{
This function transforms correlation matrices into venn diagrams. The shared surface area of circles corresponds to the shared variance (R squared) or to another metric (e.g. Pearson / spearman correlation). The algorithm is an approximation based on a quasi-Newton algorithm.
}
\examples{
cormat <- cor(mtcars[,2:13])

fit <- cor2Venn(cormat)
p<-cor2Vennplot(fit)
}
\references{
Martin Steppan (2022). corr2venn: Correlation to Venn diagramm. R package version 0.1.0.
}
\keyword{Correlation}
\keyword{plot,}
\keyword{psychometrics}
