% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormat16pf.R
\docType{data}
\name{cormat16pf}
\alias{cormat16pf}
\title{Correlations of 16 Personality Factors}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
\href{http://openpsychometrics.org/_rawdata/}{}
}
\usage{
data(cormat16pf)
}
\description{
Data from items from the International Personality Item Pool
}
\examples{
library(corrplot)
data(cormat16pf)
corrplot(cormat16pf)
}
\references{
Possible Questionnaire Format for
(\href{http://openpsychometrics.org/_rawdata/}{Link})
}
\keyword{datasets}
