% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big5.R, R/prefittedmodel.R
\docType{cor2Vennobj}
\name{prefittedmodel}
\alias{prefittedmodel}
\title{Answers to the Big Five Personality Test questionnaire (50 items)}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.

An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
This is a prefitted cor2Venn-Object for a corrrelation matrix based on answers on the Big Five Personality Test questionnaire (50 items) openly accessible from the International Personality Item Pool
\href{http://openpsychometrics.org/_rawdata/}{}

This is a prefitted cor2Venn-Object for a corrrelation matrix based on answers on the Big Five Personality Test questionnaire (50 items) openly accessible from the International Personality Item Pool
\href{http://openpsychometrics.org/_rawdata/}{}
}
\usage{
prefittedmodel

prefittedmodel
}
\description{
Data from items from the International Personality Item Pool (N=19,719)

Data from items from the International Personality Item Pool (N=19,719)
}
\examples{

cormat<- cor(big5,use="pairwise.complete.obs")


oldfit <- prefittedmodel
betterfit <- cor2Venn(cormat,Recode=TRUE,cor2dist=T,startingvalues = prefittedmodel)

cormat<- cor(big5,use="pairwise.complete.obs")


oldfit <- prefittedmodel
betterfit <- cor2Venn(cormat,Recode=TRUE,cor2dist=T,startingvalues = prefittedmodel)
}
\references{
R Package "cor2Venn"
(\href{http://openpsychometrics.org/_rawdata/}{Link})

R Package "cor2Venn"
(\href{http://openpsychometrics.org/_rawdata/}{Link})
}
\keyword{cor2Vennobj}
