% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2Venn.R
\name{cor2Venn}
\alias{cor2Venn}
\title{Correlation matrix to Venn diagram

This function transforms correlation matrices into venn diagrams. The shared surface area of circles corresponds to the shared variance (R squared) or to another metric (e.g. Pearson / spearman correlation). The algorithm is an approximation based on a quasi-Newton algorithm.}
\usage{
cor2venn(cormat, Rsquared = TRUE, cor2dist=c("0","1","2"), Recode = TRUE,
Coloring = c("PC1", "Mclust", "manual"))
}
\arguments{
\item{cormat}{A (square) correlation matrix (or other n times n numeric matrix)}

\item{Recode}{Automatically recodes the correlation matrix so that the highest correlation of each variable is always positive (defaults to TRUE)}

\item{cor2dist}{Euclidean distances between data points are proportional to correlations (positive correlation = close, negative correlation = distant). Three options are available: cor2dist="1": distances are proportional to correlations; cor2dist="2": distances are equivalent to signed R squared, cor2dist="0": no optimization of distances between nodes (except for surface overlap)}

\item{squared}{Shared surface areas between circles correspond to the squared correlation matrix (defaults to TRUE)}

\item{fillmode}{Coloring of circles. There are three options: "Eigen" = according to first eigenvector; "mclust" = according to cluster ananlysis of coordinates; "manual" = manual filling of nodes (use manualfill to provide a vector)}
}
\value{
p A ggplot2 object showing the graphical approximation

xyr A three-column matrix providing x and y coordinates for all variables in the plot and the radius as depicted in the ggplot2 object

t CPU time required to find the optimized solution
}
\description{
Correlation matrix to Venn diagram

This function transforms correlation matrices into venn diagrams. The shared surface area of circles corresponds to the shared variance (R squared) or to another metric (e.g. Pearson / spearman correlation). The algorithm is an approximation based on a quasi-Newton algorithm.
}
\examples{
cormat <- cor(mtcars[,2:13])

solve <- cor2venn(cormat)
plot(solve$p)
}
\references{
Martin Steppan (2022). corr2venn: Correlation to Venn diagramm. R package version 0.1.0.
}
\keyword{Correlation}
\keyword{plot,}
\keyword{psychometrics}
