% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2Venn.R
\name{cor2Venn}
\alias{cor2Venn}
\title{Correlation matrix to Venn diagram}
\usage{
cor2Venn(cormat, Rsquared = TRUE, cor2dist=FALSE, Recode = TRUE)
}
\arguments{
\item{cormat}{A (square) correlation matrix (or other n times n numeric matrix)}

\item{Recode}{Automatically recodes the correlation matrix so that the highest correlation of each variable is always positive (defaults to TRUE)}

\item{maxit}{The maximum number of iterations used by the optimization algorithm (optimr) defaults to 100}

\item{threshold}{Variables with an accuracy (Pearson correlation) of < threshold will be dropped from the model, and a new model is calculated using the rest of the variables (defaults to -1, i.e. no variables are dropped)}

\item{cor2dist}{Euclidean distances between data points are proportional to correlations / signed R squared (high correlation / signed R squared = close, negative correlation / low R squared = distant).}

\item{squared}{Shared surface areas between circles correspond to the squared correlation matrix (defaults to TRUE)}

\item{fillmode}{Coloring of circles. There are three options: "Eigen" = according to first eigenvector; "mclust" = according to cluster ananlysis of coordinates; "manual" = manual filling of nodes (use manualfill to provide a vector)}
}
\value{
p A ggplot2 object showing the graphical approximation

xyr A three-column matrix providing x and y coordinates for all variables in the plot and the radius as depicted in the ggplot2 object

t CPU time required to find the optimized solution
}
\description{
This function transforms correlation matrices into venn diagrams. The shared surface area of circles corresponds to the shared variance (R squared) or to another metric (e.g. Pearson / spearman correlation). The algorithm is an approximation based on a quasi-Newton algorithm.
}
\examples{
cormat <- ids2cormat

fit <- cor2Venn(cormat)
cor2Vennplot(fit)
}
\references{
Martin Steppan (2022). corr2venn: Correlation to Venn diagramm. R package version 0.1.0.
}
\keyword{Correlation}
\keyword{plot,}
\keyword{psychometrics}
